chore(team): split manager test support
